upd: tmux to 3.0a from 2.9a.
it's all so fucking tiresome, i really want to end it all. i can't be a neet or can't not be neet, having to depend on my father when he has to leave home to work is killing me. i wish i could someone could just discover me all of a sudden and make use of me. i really want to move out. i wish my t420 didn't die. life is a fuck.